html/browsers/browsing-the-web/navigating-across-documents/014.html

<!doctype html>
<link rel="help" href="https://html.spec.whatwg.org/multipage/browsing-the-web.html#navigating-across-documents">
<link rel="help" href="https://html.spec.whatwg.org/multipage/links.html#following-hyperlinks">
<link rel="help" href="https://html.spec.whatwg.org/multipage/browsing-the-web.html#javascript-protocol">
<link rel="help" href="https://html.spec.whatwg.org/multipage/text-level-semantics.html#the-a-element">
<link rel="help" href="https://html.spec.whatwg.org/multipage/webappapis.html#the-event-handler-processing-algorithm">
<link rel="help" href="https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#concept-form-submit">
<link rel="help" href="https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#submit-get-action">
<link rel="help" href="https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#plan-to-navigate">
<link rel="help" href="https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#planned-navigation">
<link rel="help" href="https://html.spec.whatwg.org/multipage/iframe-embed-object.html#the-iframe-element:about:blank">
<title>Link with onclick form submit to "javascript: url" navigation is not canceled because there is no href navigation; operations occur in exepected order.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe id="test" name="test"></iframe>
<form target="test" action="javascript:parent.onjsurl({data:location.href});"></form>
<a target="test" onclick="document.forms[0].submit();reports.push('onclick');">Test</a>
<script>
var test = async_test();
var reports = [];

// This direct sync function call is necessary to avoid
// relying on unrelated events, eg. onload, etc.
window.onjsurl = test.step_func_done(function(e) {
  // This only occurs once.
  reports.push(e.data);

  // "about:blank" is expected last, if this is not received last or
  // not received at all, the test must fail.
  if (e.data === "about:blank") {
    assert_equals(reports[0], "onclick");
    assert_equals(reports[1], "about:blank");
  } else {
    throw new Error("'javascript: url' code is expected to run asynchronously and in the target nested browsing context");
  }
});

test.step(function() {
  document.querySelector("a").click();
});
</script>

<!--
1. This test creates a form with `target="test" and an anchor with both an
    onclick and href.
2. The test clicks the anchor, which will trigger the onclick handler
    operation, which will attempt to submit the form. The form's `action` is
    a "javscript: url", which must be evaluated and executed asynchronously.
3. location.href is used to prove that the "javscript: url"'s contents are
    evaluated in the iframe.
4. If the semantics are implemented correctly, reports will be
    ["onclick", "about:blank"]

-->
